+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 2.6.2 (unreleased)
 ------------------
 
-- Bypass notBeforeDays when using the @available-slots to find the first bookable slot(used by BookinManager).
+- Bypass notBeforeDays when using the @available-slots to find the first bookable slot(used by BookingManager).
   [folix-01]
 
 - Fix day endpoint to show the available scheduling for the same day as start date.